+++ packages/melos/lib/src/commands/bootstrap.dart
@@ -162,7 +162,7 @@ mixin _BootstrapMixin on _CleanMixin {
     Package package,
   ) async {
     final pubGetArgs = ['pub', 'get'];
-    final execArgs = workspace.isFlutterWorkspace
+    final execArgs = package.isFlutterPackage
         ? ['flutter', ...pubGetArgs]
         : [if (utils.isPubSubcommand()) 'dart', ...pubGetArgs];
     final executable = currentPlatform.isWindows ? 'cmd' : '/bin/sh';
@@ -210,9 +210,8 @@ Future<void> _generateTemporaryProjects(MelosWorkspace workspace) async {
         pluginTemporaryPath,
       );
 
-      if (package.dependenciesInWorkspace.containsKey(plugin.name) ||
-          package.devDependenciesInWorkspace.containsKey(plugin.name) ||
-          package.dependencyOverridesInWorkspace.containsKey(plugin.name)) {
+      if (package.allTransitiveDependenciesInWorkspace
+          .containsKey(plugin.name)) {
         pubspec = pubspec.copy(
           dependencyOverrides: {
             ...pubspec.dependencyOverrides,