+++ packages/melos/lib/src/commands/runner.dart
@@ -18,12 +18,14 @@ import '../command_runner/version.dart';
 import '../common/changelog.dart';
 import '../common/exception.dart';
 import '../common/git.dart';
+import '../common/git_commit.dart';
 import '../common/glob.dart';
 import '../common/pending_package_update.dart';
 import '../common/platform.dart';
-import '../common/utils.dart' as utils;
 import '../common/utils.dart';
+import '../common/utils.dart' as utils;
 import '../common/versioning.dart' as versioning;
+import '../common/workspace_changelog.dart';
 import '../package.dart';
 import '../prompts/prompt.dart' as prompts;
 import '../scripts.dart';
@@ -54,19 +56,19 @@ class Melos extends _Melos
         _VersionMixin,
         _PublishMixin {
   Melos({
-    required this.workingDirectory,
+    required this.config,
     Logger? logger,
   }) : logger = logger ?? Logger.standard();
 
   @override
   final Logger logger;
   @override
-  final Directory workingDirectory;
+  final MelosWorkspaceConfig config;
 }
 
 abstract class _Melos {
   Logger get logger;
-  Directory get workingDirectory;
+  MelosWorkspaceConfig get config;
 
   Future<MelosWorkspace> createWorkspace({PackageFilter? filter}) async {
     var filterWithEnv = filter;
@@ -79,13 +81,12 @@ abstract class _Melos {
         scope: currentPlatform.environment[envKeyMelosPackages]!
             .split(',')
             .map(
-              (e) => createGlob(e, currentDirectoryPath: workingDirectory.path),
+              (e) => createGlob(e, currentDirectoryPath: config.path),
             )
             .toList(),
       );
     }
 
-    final config = await MelosWorkspaceConfig.fromDirectory(workingDirectory);
     return MelosWorkspace.fromConfig(
       config,
       filter: filterWithEnv,
@@ -111,7 +112,7 @@ abstract class _Melos {
     if (workspace.config.scripts.containsKey(scriptName)) {
       logger.stdout('Running $scriptName script...\n');
 
-      await run(scriptName: scriptName, configs: workspace.config);
+      await run(scriptName: scriptName);
     }
 
     try {
@@ -121,7 +122,7 @@ abstract class _Melos {
       if (workspace.config.scripts.containsKey(postScript)) {
         logger.stdout('Running $postScript script...\n');
 
-        await run(scriptName: postScript, configs: workspace.config);
+        await run(scriptName: postScript);
       }
     }
   }
@@ -129,6 +130,5 @@ abstract class _Melos {
   Future<void> run({
     String? scriptName,
     bool noSelect = false,
-    MelosWorkspaceConfig? configs,
   });
 }