+++ packages/melos/test/commands/bootstrap_test.dart
@@ -1,3 +1,5 @@
+import 'dart:io' as io;
+
 import 'package:melos/melos.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:pubspec/pubspec.dart';
@@ -35,9 +37,10 @@ void main() {
       );
 
       final logger = TestLogger();
+      final config = await MelosWorkspaceConfig.fromDirectory(workspaceDir);
       final melos = Melos(
         logger: logger,
-        workingDirectory: workspaceDir,
+        config: config,
       );
 
       await melos.bootstrap();
@@ -79,6 +82,23 @@ Generating IntelliJ IDE files...
       expect(aConfig.generator, 'melos');
     });
 
+    test(
+      'bootstrap transitive dependencies',
+      () => dependencyResolutionTest({
+        'a': [],
+        'b': ['a'],
+        'c': ['b'],
+      }),
+    );
+
+    test(
+      'bootstrap cyclic dependencies',
+      () => dependencyResolutionTest({
+        'a': ['b'],
+        'b': ['a'],
+      }),
+    );
+
     test('handles errors in pub get', () async {
       final workspaceDir = createTemporaryWorkspaceDirectory();
 
@@ -95,9 +115,10 @@ Generating IntelliJ IDE files...
       );
 
       final logger = TestLogger();
+      final config = await MelosWorkspaceConfig.fromDirectory(workspaceDir);
       final melos = Melos(
         logger: logger,
-        workingDirectory: workspaceDir,
+        config: config,
       );
 
       await expectLater(
@@ -131,4 +152,95 @@ e-Because a depends on package_that_does_not_exists any which doesn't exist (cou
 
     test('can supports package filter', () {}, skip: true);
   });
+}
+
+/// Tests whether dependencies are resolved correctly.
+///
+/// [packages] is a map where keys are package names and values are lists of
+/// packages names on which the package in the corresponding key depends.
+///
+/// In this example below **a** has no dependencies and **b** depends only on
+/// **a**:
+/// ```dart
+/// {
+///   'a': [],
+///   'b': ['a']
+/// }
+/// ```
+///
+/// For each entry in [packages] a package with the key as the name will be
+/// generated.
+///
+/// After running `melos bootstrap`, for each package it is verified that all
+/// direct and transitive dependencies are path dependencies with the correct
+/// path.
+Future<void> dependencyResolutionTest(
+  Map<String, List<String>> packages,
+) async {
+  final workspaceDir = createTemporaryWorkspaceDirectory();
+
+  Future<MapEntry<String, io.Directory>> createPackage(
+    MapEntry<String, List<String>> entry,
+  ) async {
+    final package = entry.key;
+    final dependencies = entry.value;
+    final directory = await createProject(
+      workspaceDir,
+      PubSpec(
+        name: package,
+        dependencies: {
+          for (final dependency in dependencies)
+            dependency: HostedReference(VersionConstraint.any),
+        },
+      ),
+    );
+
+    return MapEntry(package, directory);
+  }
+
+  final packageDirs = Map.fromEntries(
+    await Future.wait(packages.entries.map(createPackage)),
+  );
+
+  List<String> transitiveDependenciesOfPackage(String root) {
+    final transitiveDependencies = <String>[];
+    final workingSet = packages[root]!.toList();
+
+    while (workingSet.isNotEmpty) {
+      final current = workingSet.removeLast();
+
+      if (current == root) {
+        continue;
+      }
+
+      if (!transitiveDependencies.contains(current)) {
+        transitiveDependencies.add(current);
+        workingSet.addAll(packages[current]!);
+      }
+    }
+
+    return transitiveDependencies;
+  }
+
+  Future<void> validatePackage(String package) async {
+    final packageConfig = packageConfigForPackageAt(packageDirs[package]!);
+    final transitiveDependencies = transitiveDependenciesOfPackage(package);
+
+    for (final dependency in transitiveDependencies) {
+      final dependencyConfig =
+          packageConfig.packages.firstWhere((e) => e.name == dependency);
+      expect(dependencyConfig.rootUri, '../../$dependency');
+    }
+  }
+
+  final logger = TestLogger();
+  final config = await MelosWorkspaceConfig.fromDirectory(workspaceDir);
+  final melos = Melos(
+    logger: logger,
+    config: config,
+  );
+
+  await melos.bootstrap();
+
+  await Future.wait<void>(packages.keys.map(validatePackage));
 }